+++ src/main/java/lotto/domain/PurchasePrice.java
@@ -4,24 +4,24 @@ import lotto.util.ExceptionMessage;
 
 public class PurchasePrice {
 
-    private final Integer money;
+    private final Long money;
 
-    public static PurchasePrice of(Integer amount) {
+    public static PurchasePrice of(Long amount) {
         return new PurchasePrice(amount);
     }
 
-    private PurchasePrice(Integer money) {
+    private PurchasePrice(Long money) {
         validatePrice(money);
         this.money = money;
     }
 
-    private void validatePrice(Integer amount) {
+    private void validatePrice(Long amount) {
         if (amount % Lotto.PRICE != 0) {
             throw new IllegalArgumentException(ExceptionMessage.CHECK_UNIT_PRICE.getMessage());
         }
     }
 
-    public int getPurchaseLottoAmount() {
+    public Long getPurchaseLottoAmount() {
         return this.money / Lotto.PRICE;
     }
 }
